@@ -17,9 +17,9 @@
         "source": "https://github.com/php-tuf/composer-stager"
     },
     "require": {
-        "php": "^7.3 || ^8.0",
-        "symfony/filesystem": "^4.4 || ^5.2",
-        "symfony/process": "^4.4 || ^5.2"
+        "php": "^7.4 || ^8.0",
+        "symfony/filesystem": "^4.4 || ^5.2 || ^6.1",
+        "symfony/process": "^4.4 || ^5.2 || ^6.1"
     },
     "require-dev": {
         "ext-json": "*",
@@ -39,6 +39,7 @@
         "phpstan/phpstan-symfony": "^0.12.37",
         "phpunit/phpunit": "^9.5",
         "psalm/phar": "^4.8",
+        "qossmic/deptrac-shim": "^0.23.0",
         "roave/security-advisories": "dev-latest",
         "slevomat/coding-standard": "^7.0",
         "squizlabs/php_codesniffer": "^3.5",
@@ -83,10 +84,7 @@
     },
     "scripts": {
         "post-install-cmd": "phpcs --config-set installed_paths ../../phpcompatibility/php-compatibility,../../slevomat/coding-standard,phpcs/ComposerStager",
-        "all": [
-            "grumphp run",
-            "@deptrac"
-        ],
+        "all": "grumphp run",
         "c": "rm -r var/*/* &> /dev/null || true",
         "coverage": "open $(pwd)/var/phpunit/html-coverage/index.html || echo 'Run the \"coverage\" command to generate the coverage report.'",
         "deptrac": "deptrac --cache-file=var/deptrac/cache.json",
@@ -101,10 +99,7 @@
         "psalm": "psalm.phar",
         "psalter": "psalm.phar --alter --issues=all",
         "security": "@composer update --dry-run roave/security-advisories",
-        "static": [
-            "grumphp run --testsuite=static",
-            "@deptrac"
-        ],
+        "static": "grumphp run --testsuite=static",
         "test": "grumphp run --testsuite=test",
         "unit": "phpunit --testsuite=Unit"
     },
@@ -112,7 +107,7 @@
         "all": "Runs all automated tests and static analysis tools.",
         "c": "Clears the script caches in the \"var\" directory.",
         "coverage": "Opens the code coverage report generated by the \"test\" script (on macOS).",
-        "deptrac": "Enforces architectural decisions. (Assumes that qossmic/deptrac-shim is installed globally.)",
+        "deptrac": "Enforces architectural decisions.",
         "functional": "Runs functional tests.",
         "lint": "Lints PHP files.",
         "phpcbf": "Automatically fixes standards violations where possible.",
